requirements.txt
python-telegram-bot==22.1
python-dotenv==1.0.1
pydantic==2.8.2
uvicorn==0.30.6
fastapi==0.112.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
httpx==0.27.0
structlog==24.4.0
sentry-sdk==2.13.0
